+++ index.html
@@ -10,7 +10,7 @@
 
 <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
 </body>
 
 </html>
