+++ src/index.css
@@ -3,5 +3,5 @@
 @tailwind utilities;
 
 body {
-    background-color: rgb(87, 87, 87);
+  background-color: rgb(87, 87, 87);
 }
