+++ src/utils/types.ts
@@ -1,0 +1,6 @@
+export type Game = {
+    _id: string;
+    name: string;
+    rating: string;
+    cover: string;
+};
