+++ tailwind.config.js
@@ -9,4 +9,3 @@ export default {
   },
   plugins: [],
 }
-
